backend/requirements.txt
fastapi
uvicorn
PyPDF2
requests
psycopg2-binary
pgvector
python-dotenv
Pillow
pdf2image
pytesseract
# Optional: camelot for table extraction (requires ghostscript and opencv installed on the system)
# camelot-py[cv]
